+++ react/ShippingCalculator.tsx
@@ -5,6 +5,7 @@ import styles from './styles.css'
 import { useRuntime } from 'vtex.render-runtime'
 import { FormattedMessage, defineMessages } from 'react-intl'
 import { components, helpers } from 'vtex.address-form'
+import { newAddress } from './utils/address'
 const { AddressContainer, AddressRules, StyleguideInput } = components
 const { addValidation } = helpers
 
@@ -27,10 +28,10 @@ const ShippingCalculator: FunctionComponent<CustomProps> = ({
   deliveryOptions,
   countries,
 }) => {
-  const { account } = useRuntime()
+  const { account, culture } = useRuntime()
 
   const [address, setAddress] = useState<AddressWithValidation>(
-    addValidation(selectedAddress)
+    addValidation(selectedAddress || newAddress({ country: culture.country }))
   )
 
   const [showResult, setShowResult] = useState<boolean>(false)
@@ -39,8 +40,7 @@ const ShippingCalculator: FunctionComponent<CustomProps> = ({
     setAddress(address)
   }
 
-  const handleSubmit = (event: any) => {
-    event.preventDefault()
+  const handleSubmit = () => {
     const postalCodeValid =
       address && address.postalCode && address.postalCode.valid
     const geoCoordinatesValid =
@@ -52,7 +52,7 @@ const ShippingCalculator: FunctionComponent<CustomProps> = ({
   }
 
   return (
-    <div className={`${styles.container} flex flex-column pv6 ph4`}>
+    <div className={`${styles.container} flex flex-column pv6`}>
       <p className="t-heading-5 c-on-muted-3">
         <FormattedMessage id="store/shipping-calculator.delivery" />
       </p>
@@ -79,6 +79,11 @@ const ShippingCalculator: FunctionComponent<CustomProps> = ({
       </AddressRules>
     </div>
   )
+}
+
+ShippingCalculator.defaultProps = {
+  countries: [],
+  deliveryOptions: [],
 }
 
 export default ShippingCalculator